We only need a 32-bit offset into the function. This saves space on 32-bit systems.

+++ sys/sys/sdt.h
@@ -427,7 +427,7 @@ struct sdt_provider {
 struct sdt_site {
 	struct sdt_probe *sdts_probe;
 	const char 	*sdts_func;
-	uint64_t	sdts_offset;
+	uint32_t	sdts_offset;
 };
 
 SDT_PROVIDER_DECLARE(sdt);
